+++ packages/terra-framework-docs/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+* Added
+  * Added examples and tests for `terra-worklist-data-grid` to interact with cells containing focusable elements.
 
 ## 1.27.0 - (July 25, 2023)